+++ merlin/exceptions/__init__.py
@@ -105,9 +105,9 @@ class NoWorkersException(Exception):
         super().__init__(message)
 
 
-class InvalidTaskServerError(Exception):
+class MerlinInvalidTaskServerError(Exception):
     """
-    Exception to signal that the task server the user provided is not supported by Merlin.
+    Exception to signal that an invalid task server was provided.
     """
 
     def __init__(self, message):