+++ merlin/main.py
@@ -39,8 +39,10 @@ import sys
 import time
 import traceback
 from argparse import (
+    SUPPRESS,
     ArgumentDefaultsHelpFormatter,
     ArgumentParser,
+    ArgumentTypeError,
     Namespace,
     RawDescriptionHelpFormatter,
     RawTextHelpFormatter,
@@ -48,12 +50,14 @@ from argparse import (
 from contextlib import suppress
 from typing import Dict, List, Optional, Tuple, Union
 
+import yaml
 from tabulate import tabulate
 
 from merlin import VERSION, router
 from merlin.ascii_art import banner_small
 from merlin.db_scripts.db_commands import database_delete, database_get, database_info
 from merlin.db_scripts.merlin_db import MerlinDatabase
+from merlin.config.merlin_config_manager import MerlinConfigManager
 from merlin.examples.generator import list_examples, setup_example
 from merlin.log_formatter import setup_logging
 from merlin.monitor.monitor import Monitor
@@ -539,29 +543,38 @@ def print_info(args: Namespace):
 
 def config_merlin(args: Namespace):
     """
-    CLI command to set up the default Merlin configuration.
+    CLI command to manage Merlin configuration files.
 
-    This function initializes the configuration app.yaml file that's
-    necessary to connect Merlin to a central server. If the output
-    directory is not specified via the command-line arguments, it
-    defaults to the user's home directory under `.merlin`.
+    This function handles various configuration-related operations based on
+    the provided subcommand. It ensures that the specified configuration
+    file has a valid YAML extension (i.e., `.yaml` or `.yml`).
+
+    If no output file is explicitly provided, a default path is used.
 
     Args:
-        args: Parsed command-line arguments, which may include:\n
-            - `output_dir`: Path to the output directory for
-              configuration files. If not provided, defaults to
-              `~/.merlin`.
-            - `task_server`: Address of the task server for the
-              configuration.
-            - `broker`: Address of the broker service to use.
-            - `test`: Flag indicating whether to run in test mode.
+        args: Parsed command-line arguments.
     """
-    output_dir: Optional[str] = args.output_dir
-    if output_dir is None:
-        user_home: str = os.path.expanduser("~")
-        output_dir: str = os.path.join(user_home, ".merlin")
+    if args.commands != "create":  # Check that this is a valid yaml file
+        try:
+            with open(args.config_file, "r") as conf_file:
+                yaml.safe_load(conf_file)
+        except FileNotFoundError:
+            raise ArgumentTypeError(f"The file '{args.config_file}' does not exist.")
+        except yaml.YAMLError as e:
+            raise ArgumentTypeError(f"The file '{args.config_file}' is not a valid YAML file: {e}")
 
-    router.create_config(args.task_server, output_dir, args.broker, args.test)
+    config_manager = MerlinConfigManager(args)
+
+    if args.commands == "create":
+        config_manager.create_template_config()
+        config_manager.save_config_path()
+    elif args.commands == "update-broker":
+        config_manager.update_broker()
+    elif args.commands == "update-backend":
+        config_manager.update_backend()
+    elif args.commands == "use":  # Config file path is updated in constructor of MerlinConfigManager
+        config_manager.config_file = args.config_file
+        config_manager.save_config_path()
 
 
 def process_example(args: Namespace) -> None:
@@ -841,35 +854,90 @@ def setup_argparse() -> None:  # pylint: disable=R0915
         formatter_class=ArgumentDefaultsHelpFormatter,
     )
     mconfig.set_defaults(func=config_merlin)
+    # The below option makes it so the `config_path.txt` file is written to the test directory
     mconfig.add_argument(
-        "--task_server",
+        "-t",
+        "--test",
+        action="store_true",
+        help=SUPPRESS,  # Hides from `--help`
+    )
+    mconfig_subparsers = mconfig.add_subparsers(dest="commands", help="Subcommands for 'config'")
+    default_config_file = os.path.join(os.path.expanduser("~"), ".merlin", "app.yaml")
+
+    # Subcommand: melrin config create
+    config_create_parser = mconfig_subparsers.add_parser("create", help="Create a new configuration file.")
+    config_create_parser.add_argument(
+        "--task-server",
         type=str,
         default="celery",
-        help="Task server type for which to create the config.\
-                            Default: %(default)s",
+        help="Task server type for which to create the config. Default: %(default)s",
     )
-    mconfig.add_argument(
+    config_create_parser.add_argument(
         "-o",
-        "--output_dir",
+        "--output-file",
+        dest="config_file",
         type=str,
-        default=None,
-        help="Optional directory to place the default config file.\
-                            Default: ~/.merlin",
+        default=default_config_file,
+        help=f"Optional file name for your configuration. Default: {default_config_file}",
     )
-    mconfig.add_argument(
+    config_create_parser.add_argument(
         "--broker",
         type=str,
         default=None,
-        help="Optional broker type, backend will be redis\
-                            Default: rabbitmq",
+        help="Optional broker type, backend will be redis. Default: rabbitmq",
     )
-    mconfig.add_argument(
-        "--test",
-        type=str,
-        default=None,
-        help="A config used in the testing suite (or for exemplative purposes).\
-                            Default: rabbitmq",
+
+    # Subcommand: merlin config update-broker
+    config_broker_parser = mconfig_subparsers.add_parser("update-broker", help="Update broker settings in app.yaml")
+    config_broker_parser.add_argument(
+        "-t",
+        "--type",
+        required=True,
+        choices=["redis", "rabbitmq"],
+        help="Type of broker to configure (redis or rabbitmq).",
     )
+    config_broker_parser.add_argument(
+        "--cf",
+        "--config-file",
+        dest="config_file",
+        default=default_config_file,
+        help=f"The path to the config file that will be updated. Default: {default_config_file}",
+    )
+    config_broker_parser.add_argument("-u", "--username", help="Broker username (only for rabbitmq)")
+    config_broker_parser.add_argument("--pf", "--password-file", dest="password_file", help="Path to password file")
+    config_broker_parser.add_argument("-s", "--server", help="The URL of the server")
+    config_broker_parser.add_argument("-p", "--port", type=int, help="Broker port")
+    config_broker_parser.add_argument("-v", "--vhost", help="Broker vhost (only for rabbitmq)")
+    config_broker_parser.add_argument("-c", "--cert-reqs", help="Broker cert requirements")
+    config_broker_parser.add_argument("-d", "--db-num", type=int, help="Redis database number (only for redis).")
+
+    # Subcommand: merlin config update-backend
+    config_backend_parser = mconfig_subparsers.add_parser("update-backend", help="Update results backend settings in app.yaml")
+    config_backend_parser.add_argument(
+        "-t",
+        "--type",
+        required=True,
+        choices=["redis"],
+        help="Type of results backend to configure.",
+    )
+    config_backend_parser.add_argument(
+        "--cf",
+        "--config-file",
+        dest="config_file",
+        default=default_config_file,
+        help=f"The path to the config file that will be updated. Default: {default_config_file}",
+    )
+    config_backend_parser.add_argument("-u", "--username", help="Backend username")
+    config_backend_parser.add_argument("--pf", "--password-file", dest="password_file", help="Path to password file")
+    config_backend_parser.add_argument("-s", "--server", help="The URL of the server")
+    config_backend_parser.add_argument("-p", "--port", help="Backend port")
+    config_backend_parser.add_argument("-d", "--db-num", help="Backend database number")
+    config_backend_parser.add_argument("-c", "--cert-reqs", help="Backend cert requirements")
+    config_backend_parser.add_argument("-e", "--encryption-key", help="Path to encryption key file")
+
+    # Subcommand: merlin config use
+    config_use_parser = mconfig_subparsers.add_parser("use", help="Use a different configuration file.")
+    config_use_parser.add_argument("config_file", type=str, help="The path to the new configuration file to use.")
 
     # merlin example
     example: ArgumentParser = subparsers.add_parser(