+++ tests/unit/config/test_configfile.py
@@ -10,6 +10,7 @@ from copy import copy, deepcopy
 
 import pytest
 import yaml
+from pytest_mock import MockerFixture
 
 from merlin.config.configfile import (
     CONFIG,
@@ -27,12 +28,63 @@ from merlin.config.configfile import (
     set_username_and_vhost,
 )
 from tests.constants import CERT_FILES
+from tests.fixture_types import FixtureCallable, FixtureStr
 from tests.utils import create_dir
 
 
-CONFIGFILE_DIR = "{temp_output_dir}/test_configfile"
 COPIED_APP_FILENAME = "app_copy.yaml"
-DUMMY_APP_FILEPATH = f"{os.path.dirname(__file__)}/dummy_app.yaml"
+DUMMY_APP_FILEPATH = os.path.join(os.path.dirname(__file__), "dummy_app.yaml")
+
+
+@pytest.fixture(scope="session")
+def configfile_testing_dir(create_testing_dir: FixtureCallable, config_testing_dir: FixtureStr) -> FixtureStr:
+    """
+    Fixture to create a temporary output directory for tests related to testing the
+    `config` directory.
+
+    Args:
+        create_testing_dir: A fixture which returns a function that creates the testing directory.
+        config_testing_dir: The path to the temporary ouptut directory for config tests.
+
+    Returns:
+        The path to the temporary testing directory for tests of the `configfile.py` module
+    """
+    return create_testing_dir(config_testing_dir, "configfile_tests")
+
+
+@pytest.fixture(scope="session")
+def demo_app_yaml(configfile_testing_dir: FixtureStr) -> FixtureStr:
+    """
+    Fixture that creates an empty `app.yaml` file in the specified testing directory.
+
+    Args:
+        configfile_testing_dir (FixtureStr): The directory used for testing configurations.
+
+    Returns:
+        The path to the newly created `app.yaml` file.
+    """
+    app_yaml_path = os.path.join(configfile_testing_dir, "app.yaml")
+    with open(app_yaml_path, "w"):
+        pass
+    return app_yaml_path
+
+
+@pytest.fixture(scope="session")
+def config_path(configfile_testing_dir: FixtureStr, demo_app_yaml: FixtureStr) -> FixtureStr:
+    """
+    Fixture that creates a `config_path.txt` file containing the path to `app.yaml`.
+
+    Args:
+        configfile_testing_dir (FixtureStr): The directory used for testing configurations.
+        demo_app_yaml (FixtureStr): The path to the `app.yaml` file created by the `demo_app_yaml` fixture.
+
+    Returns:
+        The path to the newly created `config_path.txt` file.
+    """
+    config_path_file = os.path.join(configfile_testing_dir, "config_path.txt")
+    with open(config_path_file, "w") as cpf:
+        cpf.write(demo_app_yaml)
+    return config_path_file
 
 
 def create_app_yaml(app_yaml_filepath: str):
@@ -46,20 +98,21 @@ def create_app_yaml(app_yaml_filepath: str):
         shutil.copy(DUMMY_APP_FILEPATH, full_app_yaml_filepath)
 
 
-def test_load_config(temp_output_dir: str):
+def test_load_config(configfile_testing_dir: FixtureStr):
     """
     Test the `load_config` function.
 
-    :param temp_output_dir: The path to the temporary output directory we'll be using for this test run
+    Args:
+        configfile_testing_dir (FixtureStr): The directory used for testing configurations.
     """
-    configfile_dir = CONFIGFILE_DIR.format(temp_output_dir=temp_output_dir)
+    configfile_dir = os.path.join(configfile_testing_dir, "test_load_config")
     create_dir(configfile_dir)
     create_app_yaml(configfile_dir)
 
     with open(DUMMY_APP_FILEPATH, "r") as dummy_app_file:
         expected = yaml.load(dummy_app_file, yaml.Loader)
 
-    actual = load_config(f"{configfile_dir}/app.yaml")
+    actual = load_config(os.path.join(configfile_dir, "app.yaml"))
     assert actual == expected
 
 
@@ -70,121 +123,112 @@ def test_load_config_invalid_file():
     assert load_config("invalid/filepath") is None
 
 
-def test_find_config_file_valid_path(temp_output_dir: str):
+def test_find_config_file_config_path_file_exists_and_is_valid(
+    mocker: MockerFixture, demo_app_yaml: FixtureStr, config_path: FixtureStr
+):
     """
-    Test the `find_config_file` function with passing a valid path in.
+    Test that `find_config_file` correctly returns the path to the configuration file
+    when `CONFIG_PATH_FILE` exists and points to a valid `app.yaml` file.
 
-    :param temp_output_dir: The path to the temporary output directory we'll be using for this test run
+    Args:
+        mocker (MockerFixture): Pytest mocker fixture for mocking functionality.
+        demo_app_yaml (FixtureStr): Path to the valid `app.yaml` file.
+        config_path (FixtureStr): Path to the `CONFIG_PATH_FILE` containing the valid configuration path.
     """
-    configfile_dir = CONFIGFILE_DIR.format(temp_output_dir=temp_output_dir)
-    create_dir(configfile_dir)
-    create_app_yaml(configfile_dir)
-
-    assert find_config_file(configfile_dir) == f"{configfile_dir}/app.yaml"
+    mocker.patch("merlin.config.configfile.CONFIG_PATH_FILE", config_path)
+    result = find_config_file()
+    assert result == demo_app_yaml
 
 
-def test_find_config_file_invalid_path():
+def test_find_config_file_config_path_file_exists_but_invalid(mocker: MockerFixture, configfile_testing_dir: FixtureStr):
     """
-    Test the `find_config_file` function with passing an invalid path in.
+    Test that `find_config_file` returns `None` when `CONFIG_PATH_FILE` exists but points to an invalid path.
+
+    Args:
+        mocker (MockerFixture): Pytest mocker fixture for mocking functionality.
+        configfile_testing_dir (FixtureStr): Directory used for testing invalid configuration paths.
     """
-    assert find_config_file("invalid/path") is None
+    config_file_path = os.path.join(configfile_testing_dir, "config_path_app_yaml_doesnt_exist.txt")
+    with open(config_file_path, "w") as cfp:
+        cfp.write("invalid_app.yaml")
+    mocker.patch("merlin.config.configfile.CONFIG_PATH_FILE", config_file_path)
+    mocker.patch("merlin.config.configfile.MERLIN_HOME", os.path.join(configfile_testing_dir, ".merlin"))
+    result = find_config_file()
+    assert result is None
 
 
-def test_find_config_file_local_path(temp_output_dir: str):
+def test_find_config_file_local_app_yaml_exists(mocker: MockerFixture, configfile_testing_dir: FixtureStr):
     """
-    Test the `find_config_file` function by having it find a local (in our cwd) app.yaml file.
-    We'll use the `temp_output_dir` fixture so that our current working directory is in a temp
-    location.
+    Test that `find_config_file` correctly returns the path to `app.yaml` when it exists in the current working directory.
 
-    :param temp_output_dir: The path to the temporary output directory we'll be using for this test run
+    Args:
+        mocker (MockerFixture): Pytest mocker fixture for mocking functionality.
+        configfile_testing_dir (FixtureStr): Directory used for testing.
     """
+    mocker.patch("merlin.config.configfile.CONFIG_PATH_FILE", os.path.join(configfile_testing_dir, "invalid_config_path.txt"))
+    mocker.patch("os.getcwd", return_value=configfile_testing_dir)
+    local_app_yaml = os.path.join(configfile_testing_dir, "app.yaml")
+    mocker.patch("os.path.isfile", side_effect=lambda x: x == local_app_yaml)
 
-    # Create the configfile directory and put an app.yaml file there
-    configfile_dir = CONFIGFILE_DIR.format(temp_output_dir=temp_output_dir)
-    create_dir(configfile_dir)
-    create_app_yaml(configfile_dir)
-
-    # Move into the configfile directory and run the test
-    os.chdir(configfile_dir)
-    try:
-        assert find_config_file() == f"{os.getcwd()}/app.yaml"
-    except AssertionError as exc:
-        # Move back to the temp output directory even if the test fails
-        os.chdir(temp_output_dir)
-        raise AssertionError from exc
-
-    # Move back to the temp output directory
-    os.chdir(temp_output_dir)
+    result = find_config_file()
+    assert result == local_app_yaml
 
 
-def test_find_config_file_merlin_home_path(temp_output_dir: str):
+def test_find_config_file_merlin_home_app_yaml_exists(mocker: MockerFixture, configfile_testing_dir: FixtureStr):
     """
-    Test the `find_config_file` function by having it find an app.yaml file in our merlin directory.
-    We'll use the `temp_output_dir` fixture so that our current working directory is in a temp
-    location.
+    Test that `find_config_file` correctly returns the path to `app.yaml` when it exists in the `MERLIN_HOME` directory.
 
-    :param temp_output_dir: The path to the temporary output directory we'll be using for this test run
+    Args:
+        mocker (MockerFixture): Pytest mocker fixture for mocking functionality.
+        configfile_testing_dir (FixtureStr): Directory used for testing.
     """
-    merlin_home = os.path.expanduser("~/.merlin")
-    if not os.path.exists(merlin_home):
-        os.mkdir(merlin_home)
-    create_app_yaml(merlin_home)
-    assert find_config_file() == f"{merlin_home}/app.yaml"
+    mocker.patch("merlin.config.configfile.CONFIG_PATH_FILE", os.path.join(configfile_testing_dir, "invalid_config_path.txt"))
+    mocker.patch("merlin.config.configfile.MERLIN_HOME", configfile_testing_dir)
+    merlin_home_app_yaml = os.path.join(configfile_testing_dir, "app.yaml")
+    mocker.patch("os.path.isfile", side_effect=lambda x: x == merlin_home_app_yaml)
+
+    result = find_config_file()
+    assert result == merlin_home_app_yaml
 
 
-def check_for_and_move_app_yaml(dir_to_check: str) -> bool:
+def test_find_config_file_no_app_yaml_found(mocker: MockerFixture):
     """
-    Check for any app.yaml files in `dir_to_check`. If one is found, rename it.
-    Return True if an app.yaml was found, false otherwise.
+    Test that `find_config_file` returns `None` when no `app.yaml` file is found in any location.
 
-    :param dir_to_check: The directory to search for an app.yaml in
-    :returns: True if an app.yaml was found. False otherwise.
+    Args:
+        mocker (MockerFixture): Pytest mocker fixture for mocking functionality.
     """
-    for filename in os.listdir(dir_to_check):
-        full_path = os.path.join(dir_to_check, filename)
-        if os.path.isfile(full_path) and filename == "app.yaml":
-            os.rename(full_path, f"{dir_to_check}/{COPIED_APP_FILENAME}")
-            return True
-    return False
+    mocker.patch("os.path.isfile", return_value=False)
+    result = find_config_file()
+    assert result is None
 
 
-def test_find_config_file_no_path(temp_output_dir: str):
+def test_find_config_file_path_provided_app_yaml_exists(mocker: MockerFixture, configfile_testing_dir: FixtureStr):
     """
-    Test the `find_config_file` function by making it unable to find any app.yaml path.
-    We'll use the `temp_output_dir` fixture so that our current working directory is in a temp
-    location.
+    Test that `find_config_file` correctly returns the path to `app.yaml` when a valid directory path is provided.
 
-    :param temp_output_dir: The path to the temporary output directory we'll be using for this test run
+    Args:
+        mocker (MockerFixture): Pytest mocker fixture for mocking functionality.
+        configfile_testing_dir (FixtureStr): Directory used for testing.
     """
+    mocker.patch("merlin.config.configfile.CONFIG_PATH_FILE", os.path.join(configfile_testing_dir, "invalid_config_path.txt"))
+    mocker.patch("os.path.isfile", side_effect=lambda x: x == "/mock/provided/path/app.yaml")
+    mocker.patch("os.path.exists", return_value=True)
+    result = find_config_file("/mock/provided/path")
+    assert result == "/mock/provided/path/app.yaml"
 
-    # Rename any app.yaml in the cwd
-    cwd_path = os.getcwd()
-    cwd_had_app_yaml = check_for_and_move_app_yaml(cwd_path)
 
-    # Rename any app.yaml in the merlin home directory
-    merlin_home_dir = os.path.expanduser("~/.merlin")
-    merlin_home_had_app_yaml = check_for_and_move_app_yaml(merlin_home_dir)
+def test_find_config_file_path_provided_app_yaml_does_not_exist(mocker: MockerFixture):
+    """
+    Test that `find_config_file` returns `None` when a directory path is provided but `app.yaml` does not exist in that directory.
 
-    try:
-        assert find_config_file() is None
-    except AssertionError as exc:
-        # Reset the cwd app.yaml even if the test fails
-        if cwd_had_app_yaml:
-            os.rename(f"{cwd_path}/{COPIED_APP_FILENAME}", f"{cwd_path}/app.yaml")
-
-        # Reset the merlin home app.yaml even if the test fails
-        if merlin_home_had_app_yaml:
-            os.rename(f"{merlin_home_dir}/{COPIED_APP_FILENAME}", f"{merlin_home_dir}/app.yaml")
-
-        raise AssertionError from exc
-
-    # Reset the cwd app.yaml
-    if cwd_had_app_yaml:
-        os.rename(f"{cwd_path}/{COPIED_APP_FILENAME}", f"{cwd_path}/app.yaml")
-
-    # Reset the merlin home app.yaml
-    if merlin_home_had_app_yaml:
-        os.rename(f"{merlin_home_dir}/{COPIED_APP_FILENAME}", f"{merlin_home_dir}/app.yaml")
+    Args:
+        mocker (MockerFixture): Pytest mocker fixture for mocking functionality.
+    """
+    mocker.patch("os.path.isfile", return_value=False)
+    mocker.patch("os.path.exists", return_value=False)
+    result = find_config_file("/mock/provided/path")
+    assert result is None
 
 
 def test_set_username_and_vhost_nothing_to_load():
@@ -313,15 +357,16 @@ def test_load_defaults():
     assert actual_config == expected_config
 
 
-def test_get_config(temp_output_dir: str):
+def test_get_config(configfile_testing_dir: FixtureStr):
     """
     Test the `get_config` function.
 
-    :param temp_output_dir: The path to the temporary output directory we'll be using for this test run
+    Args:
+        configfile_testing_dir (FixtureStr): The directory used for testing configurations.
     """
 
     # Create the configfile directory and put an app.yaml file there
-    configfile_dir = CONFIGFILE_DIR.format(temp_output_dir=temp_output_dir)
+    configfile_dir = os.path.join(configfile_testing_dir, "test_get_config")
     create_dir(configfile_dir)
     create_app_yaml(configfile_dir)
 
@@ -405,60 +450,32 @@ def test_is_debug_with_merlin_debug():
         os.environ["MERLIN_DEBUG"] = debug_val
 
 
-def test_default_config_info(temp_output_dir: str):
+def test_default_config_info(mocker: MockerFixture):
     """
     Test the `default_config_info` function.
 
-    :param temp_output_dir: The path to the temporary output directory we'll be using for this test run
+    Args:
+        mocker (MockerFixture): Pytest mocker fixture for mocking functionality.
     """
+    # Mock the necessary functions/variables
+    config_file_path = "/mock/config/app.yaml"
+    debug_mode = False
+    merlin_home = "/mock/merlin_home/"
+    find_config_file_mock = mocker.patch("merlin.config.configfile.find_config_file", return_value=config_file_path)
+    is_debug_mock = mocker.patch("merlin.config.configfile.is_debug", return_value=debug_mode)
+    mocker.patch("merlin.config.configfile.MERLIN_HOME", merlin_home)
 
-    # Create the configfile directory and put an app.yaml file there
-    configfile_dir = CONFIGFILE_DIR.format(temp_output_dir=temp_output_dir)
-    create_dir(configfile_dir)
-    create_app_yaml(configfile_dir)
-    cwd = os.getcwd()
-    os.chdir(configfile_dir)
-
-    # Delete the current val of MERLIN_DEBUG and store it (if there is one)
-    reset_merlin_debug = False
-    debug_val = None
-    if "MERLIN_DEBUG" in os.environ:
-        debug_val = copy(os.environ["MERLIN_DEBUG"])
-        del os.environ["MERLIN_DEBUG"]
-        reset_merlin_debug = True
-
-    # Create the merlin home directory if it doesn't already exist
-    merlin_home = f"{os.path.expanduser('~')}/.merlin"
-    remove_merlin_home = False
-    if not os.path.exists(merlin_home):
-        os.mkdir(merlin_home)
-        remove_merlin_home = True
-
-    # Run the test
-    try:
-        expected = {
-            "config_file": f"{configfile_dir}/app.yaml",
-            "is_debug": False,
-            "merlin_home": merlin_home,
-            "merlin_home_exists": True,
-        }
-        actual = default_config_info()
-        assert actual == expected
-    except AssertionError as exc:
-        # Make sure to reset values even if the test fails
-        if reset_merlin_debug:
-            os.environ["MERLIN_DEBUG"] = debug_val
-        if remove_merlin_home:
-            os.rmdir(merlin_home)
-        raise AssertionError from exc
-
-    # Reset values if necessary
-    if reset_merlin_debug:
-        os.environ["MERLIN_DEBUG"] = debug_val
-    if remove_merlin_home:
-        os.rmdir(merlin_home)
-
-    os.chdir(cwd)
+    # Run the test and verify the output
+    expected = {
+        "config_file": config_file_path,
+        "is_debug": debug_mode,
+        "merlin_home": merlin_home,
+        "merlin_home_exists": False,
+    }
+    actual = default_config_info()
+    find_config_file_mock.assert_called_once()
+    is_debug_mock.assert_called_once()
+    assert actual == expected
 
 
 def test_get_cert_file_all_valid_args(mysql_results_backend_config: "fixture", merlin_server_dir: str):  # noqa: F821